fix:renamed from internetcomputer to icp
