+++ crates/server/src/verifywalletsignatures.rs
@@ -1,3 +1,3 @@
-pub mod internetcomputer;
+pub mod icp;
 pub mod near;
 pub mod starknet;
